add an id parameter to entsWithConfig and votedWithConfig (#419)
Currently, the only usage of `entsWithConfig` and `votedWithConfig` is in TestLeaderElectionOverwriteNewerLogs2AB. 
This test case emulates a five-peer network and spawns each peer either by `entsWithConfig` or `votedWithConfig`. 
However,  neither of them specifies an id parameter and as a result the five peers all have the same id `1`. This makes debugging and logging a little bit of creepy. 
So it'd better to add an id parameter both to `entsWithConfig` and `votedWithConfig`, and modify their usage accordingly.