Update TestProvideSnap2C in raft_test.go
`handleSnapshot` only affects states in raft, not affects states in storage. More specifically, the snapshot in memory storage will be as it is after `handleSnapshot`. As a result, when leader is willing to send a snapshot to follower, it is the empty snapshot being sent. 
However, in my implementation, such a sending will be aborted since the snapshot is empty. Therefore, this test fails.
Sending an empty snapshot is definitely shall be rejected, and I update the test case to not confuse ones like me.